+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 
 - 🧱(helm) bump chart to v0.0.6
 - 🚸(front) replace help button with a dropdown menu
+- 🔒️(back) restrict sign-in to users with an allowed OIDC role
 
 ### Fixed